- Fixed: Timidity::Channel::mono, rpn, and nrpn were not initialized. In   particular, this meant that every channel was almost certainly in mono mode,   which can sound pretty bad if the song isn't meant to be played that way. - Added bank numbers to the MIDI precaching for Timidity, since I guess I do   need to care about banks, if even the Duke MIDIs use various banks. - Fixed: snd_midiprecache only exists in Win32 builds, so gameconfigfile.cpp   shouldn't unconditionally link against it. - Fixed: pre_resample() was still disabled, and it left two samples at the end   of the new wave data uninitialized.
SVN r903 (trunk)

+++ src/sound/i_musicinterns.h
@@ -112,7 +112,7 @@ public:
 	virtual bool FakeVolume() = 0;
 	virtual bool Pause(bool paused) = 0;
 	virtual bool NeedThreadedCallback() = 0;
-	virtual void PrecacheInstruments(const BYTE *instruments, int count);
+	virtual void PrecacheInstruments(const WORD *instruments, int count);
 };
 
 // WinMM implementation of a MIDI output device -----------------------------
@@ -138,7 +138,7 @@ public:
 	bool FakeVolume();
 	bool NeedThreadedCallback();
 	bool Pause(bool paused);
-	void PrecacheInstruments(const BYTE *instruments, int count);
+	void PrecacheInstruments(const WORD *instruments, int count);
 
 protected:
 	static void CALLBACK CallbackFunc(HMIDIOUT, UINT, DWORD_PTR, DWORD, DWORD);
@@ -230,7 +230,7 @@ public:
 	bool FakeVolume();
 	bool Pause(bool paused);
 	bool NeedThreadedCallback();
-	void PrecacheInstruments(const BYTE *instruments, int count);
+	void PrecacheInstruments(const WORD *instruments, int count);
 
 protected:
 	static bool FillStream(SoundStream *stream, void *buff, int len, void *userdata);
