+++ src/sound/music_midistream.cpp
@@ -733,12 +733,13 @@ MIDIDevice::~MIDIDevice()
 // If the device can benefit from preloading the instruments, it can do so
 // now.
 //
-// For each entry, bit 7 set indicates that the instrument is percussion and
-// the lower 7 bits contain the note number to use on MIDI channel 10,
-// otherwise it is melodic and the lower 7 bits are the program number.
+// Each entry is packed as follows:
+//   Bits 0- 6: Instrument number
+//   Bits 7-13: Bank number
+//   Bit    14: Select drum set if 1, tone bank if 0
 //
 //==========================================================================
 
-void MIDIDevice::PrecacheInstruments(const BYTE *instruments, int count)
+void MIDIDevice::PrecacheInstruments(const WORD *instruments, int count)
 {
 }
